# docs/design/Endpoint/openapi-admin.yaml
openapi: 3.1.0
info:
  title: Sparr Admin Endpoints
  version: 1.0.0
  description: |
    Admin endpoint specification. Corresponds to the Admin Endpoint definition in `docs/architecture/components.md`.

    ## Authentication

    Auth routes (sign-in / sign-out / OAuth callbacks) are provided by the Auth adapter at
    `app/api/auth/[...auth]` (framework-managed via `providers.auth.handlers` delegation).
    The concrete URL scheme is adapter-defined and is not specified in this document.

    Business endpoints assume an authenticated principal via `getPrincipal()` and do not own
    sign-in or sign-out flows. `role=admin` is enforced per-route via `requireAdmin()` —
    not by individual endpoint sign-in logic.

    CSRF protection relies on the SameSite cookie policy (Auth.js-independent);
    no X-CSRF-Token header is required.

    All /api/admin/** paths require role=admin enforced per-route via requireAdmin().
    Operations touch ACCOUNT / SCENARIO / PERSONA / RESULT / SETTING via the Store.
    Player endpoints (`/api/player/**`) are documented in `./openapi.yaml`.
servers:
  - url: http://localhost:3000
    description: Local dev
...
paths:
...
  /api/admin/users/{id}:
    get:
      tags: [Users]
      summary: Get user detail (including result history)
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Call Store.accounts.get(id) (404 if not found)
        4. Call Store.results.listByAccount(id) and Store.scenarios.list(),
           then join each result with its scenario to attach scenarioTitle
        5. Return AdminUserDetail (profile + results[]) with 200

      operationId: getAdminUserById
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User account ID
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Store retrieval failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Users]
      summary: Update user info
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Accept AdminUserUpdateRequest (username? / role?)
        4. If attempting to change own role, return 400 (lockout prevention)
        5. Call Store.accounts.update(id, fields) (404 if not found)
        6. Return updated AdminUserSummary with 200
      operationId: patchAdminUserById

      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User account ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdateRequest"
      responses:
        "200":
          description: Updated user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserSummary"
        "400":
          description: Validation error or attempt to change own role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Update failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Users]
      summary: Delete a user
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. If attempting to delete own account, return 400 (lockout prevention)
        4. Call Store.accounts.delete(id) (404 if not found)
        5. Return 204

      operationId: deleteAdminUserById
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User account ID
      responses:
        "204":
          description: Deletion successful
        "400":
          description: Attempt to delete own account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Deletion failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Tickets ────────────────────────────────────────────
  # Manage a player's ticket ledger. All operations target role=player accounts.
...
  /api/admin/users/{id}/tickets/{ticketId}:
    delete:
      tags: [Tickets]
      summary: Revoke a single active ticket
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Look up the target account (404 "not_found"); 400 "not_a_player" if not a player
        4. Locate the ticket by ticketId; 404 "not_found" if it does not exist or belongs to a different user
        5. Return 400 "already_consumed" if the ticket is no longer active
        6. Logically revoke the ticket (revokedAt set, isActive false) and return the updated TicketSnapshot with 200
      operationId: deleteAdminUserTicketById
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User account ID
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
          description: Ticket ledger row ID
      responses:
        "200":
          description: Ticket revoked; updated snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketSnapshot"
        "400":
          description: Ticket already inactive (error "already_consumed") or account is not a player (error "not_a_player")
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found, or ticket not found / owned by a different user (error "not_found")
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Revoke failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Scenarios ──────────────────────────────────────────
  # CRUD on SCENARIO + PERSONA (the data model follows docs/architecture/data-model.md).
...
  /api/admin/settings:
    get:
      tags: [Settings]
      summary: Get all system settings
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Call Store.settings.list()
        4. Return SystemSetting[] with 200

      operationId: getAdminSettings
      security:
        - SessionCookie: []
      responses:
        "200":
          description: System settings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SystemSetting"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Store retrieval failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Settings]
      summary: Update system settings
      description: |
        1. Verify session via SessionCookie (401 if unauthenticated)
        2. Verify role=admin (403 if not admin)
        3. Accept SettingsUpdateRequest (settings[] or key/value map)
        4. Call Store.settings.update(key, value) for each entry (upsert)
        5. Return updated SystemSetting[] with 200

        - Keys not included in the request are left unchanged (partial update)
        - Keys that do not exist yet are created (upsert)
      operationId: postAdminSettings

      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdateRequest"
      responses:
        "200":
          description: All system settings after update
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SystemSetting"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Update failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: |
        Session cookie set by the Auth adapter on sign-in. Sent automatically by the browser
        for same-origin requests. CSRF protection relies on the SameSite cookie policy —
        no separate X-CSRF-Token header is required.
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth / Account ────────────────────────────────────
...
    RecentResultSummary:
      type: object
      required: [id, scenarioId, score, headline, createdAt]
      properties:
        id:
          type: string
        accountId:
          type: [string, "null"]
        accountEmail:
          type: [string, "null"]
          description: |
            Email address from a joined account lookup. "null" for anonymous plays
            (accountId is "null") or a deleted account. UI fallback label when
            accountUsername is "null": accountEmail, then "—".
        accountUsername:
          type: [string, "null"]
          description: |
            Display name from a joined account lookup. "null" if the account has
            no username set (blanked out via admin edit), or for anonymous plays
            / a deleted account.

        scenarioId:
          type: string
        scenarioTitle:
          type: string
          description: Scenario title from a joined scenario lookup (falls back to scenarioId if the title is unset)
        score:
          type: number
        headline:
          type: string
        createdAt:
          type: string
...
    AdminUserSummary:
      type: object
      required: [id, email, role, username]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        role:
          type: string
          enum: [player, admin]
    AdminUserDetail:
      type: object
      required: [id, email, role, username, results]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        role:
          type: string
          enum: [player, admin]
        results:
          type: array
          items:
            $ref: "#/components/schemas/RecentResultSummary"
          description: Full result history for this user
...
    AdminUserUpdateRequest:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
          enum: [player, admin]
          description: Cannot change own role (returns 400)

    # ── Scenarios ─────────────────────────────────────────
    # Mirrors docs/architecture/data-model.md: SCENARIO + PERSONA tables.
...
    SystemSetting:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
          description: "Setting key (e.g. AI_PROVIDER / VOICE_PROVIDER)"
        value:
          type: string
          description: Setting value
    SettingsUpdateRequest:
      type: object
      required: [settings]
      properties:
        settings:
          oneOf:
            - type: array
              items:
                $ref: "#/components/schemas/SystemSetting"
              minItems: 1
            - type: object
              additionalProperties:
                type: string
          description: Settings entries to update (array or key/value map); keys not included are left unchanged

    # ── Tickets ───────────────────────────────────────────
    TicketLedger:
      type: object
      required:
        [
          id,
          accountId,
          type,
          isActive,
          consumedAt,
          consumedScenarioId,
          revokedAt,
          createdAt,
        ]
      properties:
        id:
          type: string
        accountId:
          type: string
        type:
          type: string
          enum: [registration_grant, monthly_grant, purchase, admin_adjust]
        isActive:
          type: boolean
        consumedAt:
          type: [string, "null"]
          description: ISO timestamp when the ticket was consumed, or null if unused
        consumedScenarioId:
          type: [string, "null"]
          description: Scenario id the ticket was spent on, or null if unused
        revokedAt:
          type: [string, "null"]
          description: ISO timestamp when the ticket was revoked by an admin, or null
        createdAt:
          type: string
    TicketSnapshot:
      type: object
      required: [balance, nextGrantAt, ledger]
      properties:
        balance:
          type: integer
          description: Count of currently active (unconsumed, non-revoked) tickets
        nextGrantAt:
          type: string
          description: ISO timestamp of the next scheduled grant
        ledger:
          type: array
          items:
            $ref: "#/components/schemas/TicketLedger"
          description: Up to the 20 most recent ticket rows
...